application/requirements.txt
pandas
google-cloud-bigquery
streamlit
requests